Crate ethcore_light

Light client logic and implementation.

A "light" client stores very little chain-related data locally unlike a full node, which stores all blocks, headers, receipts, and more.

This enables the client to have a much lower resource footprint in exchange for the cost of having to ask the network for state data while responding to queries. This makes a light client unsuitable for low-latency applications, but perfectly suitable for simple everyday use-cases like sending transactions from a personal account.

It starts by performing a header-only sync, verifying random samples of members of the chain to varying degrees.

Reexports

pub use self::provider::Provider;

Modules

client
Light client implementation. Stores data from light sync

net
LES Protocol Version 1 implementation.

provider
A provider for the LES protocol. This is typically a full node, who can give as much data as necessary to its peers.

request
LES request types.