Module ethcore_light::cht

Canonical hash trie definitions and helper functions.

Constants

SIZE
The size of each CHT.

Functions

block_to_cht_number
Convert a block number to a CHT number. Returns None for block_num == 0, Some otherwise.

start_number
Get the starting block of a given CHT. CHT 0 includes block 1...SIZE, CHT 1 includes block SIZE + 1 ... 2*SIZE More generally: CHT N includes block (1 + N*SIZE)...((N+1)*SIZE). This is because the genesis hash is assumed to be known and including it would be redundant.