Module ethcore_light::net

LES Protocol Version 1 implementation.

This uses a "Provider" to answer requests. See https://github.com/ethcore/parity/wiki/Light-Ethereum-Subprotocol-(LES)

Structs

Announcement
An announcement of new chain head or capabilities made by a peer.

Capabilities
Peer capabilities.

LightProtocol
This is an implementation of the light ethereum network protocol, abstracted over a Provider of data and a p2p network.

Params
Protocol parameters.

ReqId
A request id.

Status
A peer status message.

Constants

MAX_PROTOCOL_VERSION
Max protocol version.

PACKET_COUNT
Packet count for LES.

PROTOCOL_VERSIONS
Supported protocol versions.

Traits

EventContext
Context for a protocol event.

Handler
An LES event handler.

IoContext
An I/O context which allows sending and receiving packets as well as disconnecting peers. This is used as a generalization of the portions of a p2p network which the light protocol structure makes use of.