Module ethcore_light::net::buffer_flow

LES buffer flow management.

Every request in the LES protocol leads to a reduction of the requester's buffer value as a rate-limiting mechanism. This buffer value will recharge at a set rate.

This module provides an interface for configuration of buffer flow costs and recharge rates.

Current default costs are picked completely arbitrarily, not based on any empirical timings or mathematical models.

Structs

Buffer
Buffer value.

Cost
A request cost specification.

CostTable
A cost table, mapping requests to base and per-request costs.

FlowParams
A buffer-flow manager handles costs, recharge, limits