Struct ethcore_light::net::LightProtocol

pub struct LightProtocol { /* fields omitted */ }

This is an implementation of the light ethereum network protocol, abstracted over a Provider of data and a p2p network.

This is simply designed for request-response purposes. Higher level uses of the protocol, such as synchronization, will function as wrappers around this system.

Methods

impl LightProtocol

fn new(provider: Arc<Provider>, params: Params) -> Self

Create a new instance of the protocol manager.

fn max_requests(&self, peer: PeerId, kind: Kind) -> Option<usize>

Check the maximum amount of requests of a specific type which a peer would be able to serve.

fn request_from(&self,
io: &IoContext,
peer_id: &PeerId,
request: Request)
-> Result<ReqId, Error>

Make a request to a peer.

Fails on: nonexistent peer, network error, peer not server, insufficient buffer. Does not check capabilities before sending. On success, returns a request id which can later be coordinated with an event.

fn make_announcement(&self, io: &IoContext, announcement: Announcement)

Make an announcement of new chain head and capabilities to all peers. The announcement is expected to be valid.

fn add_handler(&mut self, handler: Box<Handler>)

Add an event handler. Ownership will be transferred to the protocol structure, and the handler will be kept alive as long as it is. These are intended to be added when the protocol structure is initialized as a means of customizing its behavior.

fn abort(&self)

Signal to handlers that network activity is being aborted and clear peer data.