Trait ethcore_light::net::Handler

pub trait Handler: Send + Sync {
    fn on_connect(&self,
              _ctx: &EventContext,
              _status: &Status,
              _capabilities: &Capabilities) { ... }
    fn on_disconnect(&self, _ctx: &EventContext, _unfulfilled: &[ReqId]) { ... }
    fn on_announcement(&self, _ctx: &EventContext, _announcement: &Announcement) { ... }
    fn on_transactions(&self, _ctx: &EventContext, _relay: &[SignedTransaction]) { ... }
    fn on_block_bodies(&self,
                   _ctx: &EventContext,
                   _req_id: ReqId,
                   _bodies: &[Bytes]) { ... }
    fn on_block_headers(&self,
                    _ctx: &EventContext,
                    _req_id: ReqId,
                    _headers: &[Bytes]) { ... }
    fn on_receipts(&self,
               _ctx: &EventContext,
               _req_id: ReqId,
               _receipts: &[Vec<Receipt>]) { ... }
    fn on_state_proofs(&self,
                   _ctx: &EventContext,
                   _req_id: ReqId,
                   _proofs: &[Vec<Bytes>]) { ... }
    fn on_code(&self, _ctx: &EventContext, _req_id: ReqId, _codes: &[Bytes]) { ... }
    fn on_header_proofs(&self,
                    _ctx: &EventContext,
                    _req_id: ReqId,
                    _proofs: &[(Bytes, Vec<Bytes>)]) { ... }
    fn on_abort(&self) { ... }
}

An LES event handler.

Each handler function takes a context which describes the relevant peer and gives references to the IO layer and protocol structure so new messages can be dispatched immediately.

Request responses are not guaranteed to be complete or valid, but passed IDs will be correct. Response handlers are not given a copy of the original request; it is assumed that relevant data will be stored by interested handlers.

Provided Methods

fn on_connect(&self,
_ctx: &EventContext,
_status: &Status,
_capabilities: &Capabilities)

Called when a peer connects.

fn on_disconnect(&self, _ctx: &EventContext, _unfulfilled: &[ReqId])

Called when a peer disconnects, with a list of unfulfilled request IDs as of yet.

fn on_announcement(&self, _ctx: &EventContext, _announcement: &Announcement)

Called when a peer makes an announcement.

fn on_transactions(&self, _ctx: &EventContext, _relay: &[SignedTransaction])

Called when a peer requests relay of some transactions.

fn on_block_bodies(&self,
_ctx: &EventContext,
_req_id: ReqId,
_bodies: &[Bytes])

Called when a peer responds with block bodies.

fn on_block_headers(&self,
_ctx: &EventContext,
_req_id: ReqId,
_headers: &[Bytes])

Called when a peer responds with block headers.

fn on_receipts(&self,
_ctx: &EventContext,
_req_id: ReqId,
_receipts: &[Vec<Receipt>])

Called when a peer responds with block receipts.

fn on_state_proofs(&self,
_ctx: &EventContext,
_req_id: ReqId,
_proofs: &[Vec<Bytes>])

Called when a peer responds with state proofs. Each proof is a series of trie nodes in ascending order by distance from the root.

fn on_code(&self, _ctx: &EventContext, _req_id: ReqId, _codes: &[Bytes])

Called when a peer responds with contract code.

fn on_header_proofs(&self,
_ctx: &EventContext,
_req_id: ReqId,
_proofs: &[(Bytes, Vec<Bytes>)])

Called when a peer responds with header proofs. Each proof is a block header coupled with a series of trie nodes is ascending order by distance from the root.

fn on_abort(&self)

Called on abort.