Module ethcore_light::on_demand

On-demand chain requests over LES. This is a major building block for RPCs. The request service is implemented using Futures. Higher level request handlers will take the raw data received here and extract meaningful results from it.

Modules

request
Request types, verification, and verification errors.

Structs

OnDemand
On demand request service. See module docs for more details. Accumulates info about all peers' capabilities and dispatches requests to them accordingly.

Response
Future which awaits a response to an on-demand request.

Enums

Error
Errors which can occur while trying to fulfill a request.