Struct ethcore_light::on_demand::OnDemand

pub struct OnDemand { /* fields omitted */ }

On demand request service. See module docs for more details. Accumulates info about all peers' capabilities and dispatches requests to them accordingly.

Methods

impl OnDemand

fn header_by_number(&self,
ctx: &BasicContext,
req: HeaderByNumber)
-> Response<Header>

Request a header by block number and CHT root hash.

fn header_by_hash(&self,
ctx: &BasicContext,
req: HeaderByHash)
-> Response<Header>

Request a header by hash. This is less accurate than by-number because we don't know where in the chain this header lies, and therefore can't find a peer who is supposed to have it as easily.

fn block(&self, ctx: &BasicContext, req: Body) -> Response<Block>

Request a block, given its header. Block bodies are requestable by hash only, and the header is required anyway to verify and complete the block body -- this just doesn't obscure the network query.

fn block_receipts(&self,
ctx: &BasicContext,
req: BlockReceipts)
-> Response<Vec<Receipt>>

Request the receipts for a block. The header serves two purposes: provide the block hash to fetch receipts for, and for verification of the receipts root.

fn account(&self, ctx: &BasicContext, req: Account) -> Response<BasicAccount>

Request an account by address and block header -- which gives a hash to query and a state root to verify against.