Module ethcore_light::provider

A provider for the LES protocol. This is typically a full node, who can give as much data as necessary to its peers.

Traits

Provider
Defines the operations that a provider for LES must fulfill.