Module ethcore_light::request

LES request types.

Structs

Bodies
A request for specific block bodies.

ContractCode
A request for contract code.

ContractCodes
A request for contract code.

HeaderProof
A request for a header proof from the Canonical Hash Trie.

HeaderProofs
A request for header proofs from the CHT.

Headers
A request for block headers.

Receipts
A request for transaction receipts.

StateProof
A request for a state proof

StateProofs
A request for state proofs.

Enums

HashOrNumber
Either a hash or a number.

Kind
Kinds of requests.

Request
Encompasses all possible types of requests in a single structure.